requirements.txt
pillow
requests
pyperclip